Keep the documentation link up-to-date
This trivial change will help keep the documentation link up to date.

Thanks for this nicely documented contribution. A Gradle plugin is something that many users have been asking for, but we haven't been able to prioritise our efforts towards it, so far. I will happily point them to your implementation in the near future!