Run jooq in a forked process
The jooq generator is no longer determined by the jooq version that
this plugin was compiled against. Instead, the user explictly configures
the jOOQ version and edition and the plugin uses the correct generator
based on that.

The jOOQ classpath is now a normal configuration instead of using
the buildscript classpath. This makes usage of the `plugins` DSL possible.
It also opens up the possibility to easily extend the generator classpath,
e.g. with custom generator extensions compiled inside the same project.

The version of jooq that this plugin compiles against is only used
to get a safe DSL for writing the jooq configuration. Since the
configuration XML format is pretty stable, this will make it much less
problematic if the plugin lags behind a few minor releases from time to time.