Stabilise the 'configuration' input for the task across locations and JVMs
The previous strategy of using a deep hash code had two problems:

Hash codes of enum values are only stable for a given JVM as they are based on the memory location of the Enum class. Furthermore, object hash codes are by design not very collision resistant (i.e. can produce false equality when used this way). Instead, now we use the actual bytes of the generated config file as the input, which is stable.

Secondly, by default the configuration contains an absolute path for the destination which prevent portability. We now relativise this to the project directory, which is an assumption that was already being made by getOutputDirectory(). To ensure correctness, the working directory for the jooq process is now _forced_ to be the project directory, regardless of what the user supplied exec spec configurer dictates.

There is another minor change in here to use a more conventional temp location to write the config file.