fix: lazy configuration not possible
Its not possible to use lazy configurations using things like NamedDomainObjectContainer#register because the collection is never evaluated! While configureEach is lazy within itself. The collection is never queried so it is never even instantiated because why would it - nobody ready it.

This problem is currently not solvable without forcing users to access `generateJooq` task without wrapping it in `afterEvaluate`!

This can only be solved without a breaking change if https://github.com/gradle/gradle/issues/26881 is resolved.

REF: https://github.com/gradle/gradle/issues/26881
Signed-off-by: Nils Brugger <nils.brugger@willhaben.at>